# openapi/hera.v1.yaml
openapi: 3.0.3
info:
  title: HERA Enterprise API
  description: |
    HERA (Hierarchical Enterprise Resource Architecture) API provides universal ERP capabilities
    through a revolutionary 6-table architecture. All operations maintain strict multi-tenant
    isolation and smart code integration.
  version: 1.0.0
  contact:
    name: HERA Support
    email: support@heraerp.com
  license:
    name: Proprietary
    url: https://heraerp.com/license
servers:
  - url: https://api.heraerp.com/api/v1
    description: Production
  - url: https://api.staging.heraerp.com/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /reports/{type}:
    get:
      operationId: generateReport
      summary: Generate financial report
      tags: [Reports]
      x-smart-code-families:
        - HERA.FIN.REPORT.**
      parameters:
        - $ref: '#/components/parameters/ReportType'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/Period'
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: currency
          in: query
          schema:
            type: string
            default: USD
        - name: include_details
          in: query
          schema:
            type: boolean
            default: false
        - name: comparison_period
          in: query
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/Report'
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/RateLimitExceeded'
...
  /guardrails/validate:
    post:
      operationId: validatePayload
      summary: Validate payload (dry-run)
      tags: [Guardrails]
      x-smart-code-families:
        - HERA.GUARDRAIL.VALIDATE.**
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardrailValidation'
      responses:
        200:
          $ref: '#/components/responses/GuardrailResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    OrganizationId:
      name: organization_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Organization context for multi-tenant isolation
...
    ReportType:
      name: type
      in: path
      required: true
      schema:
        type: string
        enum: [pl, balance_sheet, cashflow, trial_balance]
    Period:
      name: period
      in: query
      required: true
      schema:
        type: string
        pattern: '^(\\d{4}|\\d{4}-Q[1-4]|\\d{4}-(0[1-9]|1[0-2]))$'
...
  schemas:
    Error:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        guardrail_id:
          type: string
        autofix_applied:
          type: boolean
        ai_confidence:
          type: number
...
    Report:
      type: object
      properties:
        report_type:
          type: string
        organization_id:
          type: string
        period:
          type: string
        generated_at:
          type: string
          format: date-time
        currency:
          type: string
        sections:
          type: array
          items:
            type: object
        totals:
          type: object
          additionalProperties:
            type: number
...
    GuardrailValidation:
      type: object
      required: [table, payload]
      properties:
        table:
          type: string
        payload:
          type: object
        organization_id:
          type: string
  responses:
...
    Report:
      description: Financial report
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              report:
                $ref: '#/components/schemas/Report'
              comparison:
                type: object
              metadata:
                type: object
...
    GuardrailResult:
      description: Guardrail validation result
      content:
        application/json:
          schema:
            type: object
            properties:
              valid:
                type: boolean
              fixes_applied:
                type: integer
              corrected_payload:
                type: object
              errors:
                type: array
                items:
                  type: object
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'